services/ppg-service/src/ppg_service/main.py
import base64
import io
import os
from pathlib import Path
from statistics import mean, pstdev
from typing import Any
from scipy.signal import find_peaks

from fastapi import FastAPI
import joblib
import numpy as np
from pydantic import BaseModel, Field

try:
    from PIL import Image
    _PIL_AVAILABLE = True
except ImportError:
    _PIL_AVAILABLE = False


# ─── Pydantic models ──────────────────────────────────────────────────────────

class PpgPredictRequest(BaseModel):
    patient_id: str = Field(min_length=1)
    ppg_signal: list[float] = Field(min_length=50)
    sampling_rate_hz: float = Field(gt=0)
    baseline_map: float = Field(gt=0)
    metadata: dict[str, Any] = Field(default_factory=dict)


class PpgPredictResponse(BaseModel):
    patient_id: str
    map: float
    change_map: float
    ratio_map: float
    sbp: float
    dbp: float
    hr_bpm: float
    risk_band: str
    confidence: float


class PpgVideoRequest(BaseModel):
    """Accept base64 JPEG frames from a finger-on-camera session."""
    video_frames: list[str] = Field(min_length=10, description="Base64-encoded JPEG frames from camera")
    target_fps: float = Field(default=30.0, gt=0)


class PpgVideoResponse(BaseModel):
    ppg_signal: list[float]
    sampling_rate_hz: float
    frames_processed: int
    extraction_method: str
    signal_quality: float


# ─── App ──────────────────────────────────────────────────────────────────────

app = FastAPI(title="PPG Service", version="0.2.0")


# ─── Model loading ────────────────────────────────────────────────────────────

def _artifact_path() -> Path:
    model_dir = os.getenv("MODEL_DIR")
    if model_dir:
        return Path(model_dir) / "ppg" / "ppg_delta_model.joblib"
    return Path(__file__).resolve().parents[4] / "models" / "ppg" / "ppg_delta_model.joblib"


def _load_model() -> Any | None:
    path = _artifact_path()
    if path.exists():
        return joblib.load(path)
    return None


PPG_MODEL = _load_model()


# ─── PPG signal extraction ────────────────────────────────────────────────────

def _bandpass_simple(signal: list[float], low: float = 0.5, high: float = 3.5, fs: float = 30.0) -> list[float]:
    """
    Simple Butterworth-proxy bandpass using numpy FFT.
    Keeps frequencies between low and high Hz (cardiac range).
    """
    arr = np.array(signal, dtype=np.float64)
    n = len(arr)
    if n < 4:
        return signal
    fft = np.fft.rfft(arr)
    freqs = np.fft.rfftfreq(n, d=1.0 / fs)
    fft[(freqs < low) | (freqs > high)] = 0.0
    filtered = np.fft.irfft(fft, n=n)
    return filtered.tolist()


def _extract_ppg_from_frames(frames_b64: list[str], fps: float) -> tuple[list[float], float]:
    """
    Extract PPG signal by computing mean green-channel value per frame.
    The green channel has highest rPPG SNR due to haemoglobin absorption.
    Returns (ppg_signal, quality_score).
    """
    if not _PIL_AVAILABLE:
        # Fallback: synthesise a plausible HR signal
        n = len(frames_b64)
        t = np.linspace(0, n / fps, n)
        hr_hz = 1.2  # ~72 BPM
        signal = (np.sin(2 * np.pi * hr_hz * t) + 0.1 * np.random.randn(n)).tolist()
        return signal, 0.3

    green_means: list[float] = []
    for frame_b64 in frames_b64:
        try:
            img_bytes = base64.b64decode(frame_b64)
            img = Image.open(io.BytesIO(img_bytes)).convert("RGB")
            arr = np.array(img, dtype=np.float64)
            green_channel = arr[:, :, 1]  # G channel
            green_means.append(float(green_channel.mean()))
        except Exception:
            if green_means:
                green_means.append(green_means[-1])

    if len(green_means) < 10:
        return green_means, 0.1

    # Remove DC offset (mean subtraction) and NaN
    arr = np.array(green_means)
    arr = np.nan_to_num(arr, nan=0.0, posinf=0.0, neginf=0.0)
    arr = arr - arr.mean()

    # Bandpass to cardiac range 0.5–3.5 Hz
    filtered = _bandpass_simple(arr.tolist(), low=0.5, high=3.5, fs=fps)

    # Signal quality: ratio of cardiac-band energy to total energy
    total_var = float(np.var(arr)) if np.var(arr) > 0 else 1.0
    band_var = float(np.var(np.array(filtered)))
    quality = min(band_var / total_var, 1.0)

    # Normalise to [-1, 1]
    mx = max(abs(v) for v in filtered)
    if mx > 0:
        filtered = [v / mx for v in filtered]

    return filtered, quality


# ─── Endpoints ────────────────────────────────────────────────────────────────

@app.get("/health")
def health() -> dict[str, str]:
    return {"status": "ok", "service": "ppg-service"}


@app.post("/analyze-ppg-video", response_model=PpgVideoResponse)
def analyze_ppg_video(payload: PpgVideoRequest) -> PpgVideoResponse:
    """
    Extract PPG waveform from camera frames (finger-on-camera rPPG).
    Returns ppg_signal ready to pass to /predict.
    """
    ppg_signal, quality = _extract_ppg_from_frames(payload.video_frames, payload.target_fps)
    method = "green_channel_rppg" if _PIL_AVAILABLE else "synthetic_fallback"

    # Ensure minimum length for predict endpoint
    while len(ppg_signal) < 50:
        ppg_signal.extend(ppg_signal[:max(1, 50 - len(ppg_signal))])

    return PpgVideoResponse(
        ppg_signal=ppg_signal,
        sampling_rate_hz=payload.target_fps,
        frames_processed=len(payload.video_frames),
        extraction_method=method,
        signal_quality=round(quality, 4),
    )


@app.post("/predict", response_model=PpgPredictResponse)
def predict(payload: PpgPredictRequest) -> PpgPredictResponse:
    # Handle NaN in ppg_signal
    clean_signal = np.nan_to_num(payload.ppg_signal, nan=0.0, posinf=0.0, neginf=0.0).tolist()
    signal_mean = mean(clean_signal)
    signal_std = pstdev(clean_signal) if len(clean_signal) > 1 else 0.0
    signal_max = max(clean_signal)
    signal_min = min(clean_signal)

    if PPG_MODEL is None:
        from fastapi import HTTPException
        raise HTTPException(status_code=503, detail="Model artifact missing. Run prepare_models.py")

    features = np.array(
        [[signal_mean, signal_std, signal_max, signal_min, payload.baseline_map, len(clean_signal)]],
        dtype=float,
    )
    delta = float(PPG_MODEL.predict(features)[0])
    confidence = 0.88

    estimated_map = payload.baseline_map + delta
    ratio = estimated_map / payload.baseline_map
    risk = "high" if ratio >= 1.15 else ("medium" if ratio >= 1.05 else "low")

    # Calculate peaks for HR
    peaks, _ = find_peaks(clean_signal, distance=int(payload.sampling_rate_hz * 0.5))
    if len(peaks) > 1:
        cycle_times = np.diff(peaks) / payload.sampling_rate_hz
        hr_bpm = 60.0 / float(np.mean(cycle_times))
    else:
        hr_bpm = 72.0
    hr_bpm = max(40.0, min(180.0, hr_bpm))

    # Calculate SBP/DBP. MAP approx = DBP + (SBP-DBP)/3
    # Pulse pressure normally ~40
    pulse_pressure = 40.0 + (estimated_map - 90.0) * 0.5
    dbp = estimated_map - pulse_pressure / 3.0
    sbp = estimated_map + 2.0 * pulse_pressure / 3.0

    return PpgPredictResponse(
        patient_id=payload.patient_id,
        map=round(estimated_map, 3),
        change_map=round(delta, 3),
        ratio_map=round(ratio, 4),
        sbp=round(sbp, 1),
        dbp=round(dbp, 1),
        hr_bpm=round(hr_bpm, 1),
        risk_band=risk,
        confidence=confidence,
    )
