requirements-backend.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
numpy==2.2.6
scikit-learn==1.7.0
joblib==1.5.1
pillow==11.2.1
opencv-python-headless==4.10.0.84
httpx==0.28.1
transformers
torch
accelerate
